pyproject.toml
[tool.poetry]
name = "pptxlate"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
click = "7.1.2"
python-pptx = "^0.6.23"
tqdm = "4.58.0"
pywin32 = "300"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
